Correcciones clase Producto
- Se cambia funciones por expresiones lambda
- Se crean constantes de strings repetidos
- Se empaquetan las clases
- Se instancian clases de swing en sus respectivos métodos
- Se crea función exceptionManager para manejar las excepciones
- Se reemplaza las veces que se crea el formato de fechas para los logs por una función llamado obtenerFecha
- Se maneja el uso de los archivos dentro del try
