+++ PuntoDeVenta.iml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="SonarLintModuleSettings">
+    <option name="uniqueId" value="d884858e-8b80-4168-a491-690720693b8c" />
+  </component>
+</module>
