+++ src/main/java/com/aplicaion/Conexion.java
@@ -1,3 +1,4 @@
+package com.aplicaion;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
