+++ src/main/java/com/aplicaion/PantallaVenta.java
@@ -1,3 +1,4 @@
+package com.aplicaion;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -11,7 +12,6 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Calendar;
 import java.util.GregorianCalendar;
-import java.util.Locale;
 import javax.swing.JOptionPane;
 //import static producto.getConnection;
 //import static producto.getConnection;
